requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e file:.#lizard-auth-server  # via -r requirements.in
appdirs==1.4.4            # via black
black==20.8b1             # via -r requirements.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black
coverage==5.3             # via coveralls, lizard-auth-server
coveralls==2.1.2          # via -r requirements.in
dataclasses==0.7          # via black
django-appconf==1.0.4     # via lizard-auth-server
django-extensions==3.0.9  # via lizard-auth-server
django-nose==1.4.7        # via lizard-auth-server
django-oidc-provider==0.7.0  # via lizard-auth-server
django==1.11.29           # via django-appconf, lizard-auth-server
docopt==0.6.2             # via coveralls
factory-boy==3.1.0        # via lizard-auth-server
faker==4.14.0             # via factory-boy, lizard-auth-server
flake8==3.8.4             # via -r requirements.in
future==0.18.2            # via pyjwkest
idna==2.10                # via requests
importlib-metadata==2.0.0  # via flake8
isort==5.6.4              # via -r requirements.in
itsdangerous==1.1.0       # via lizard-auth-server
mccabe==0.6.1             # via flake8
mock==4.0.2               # via lizard-auth-server
mypy-extensions==0.4.3    # via black
nose==1.3.7               # via django-nose, lizard-auth-server
pathspec==0.8.0           # via black
psycopg2-binary==2.8.6    # via lizard-auth-server
pycodestyle==2.6.0        # via flake8
pycryptodomex==3.9.8      # via pyjwkest
pyflakes==2.2.0           # via flake8
pyjwkest==1.4.2           # via django-oidc-provider
pyjwt==1.7.1              # via lizard-auth-server
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django, lizard-auth-server
regex==2020.10.15         # via black
requests==2.24.0          # via coveralls, lizard-auth-server, pyjwkest
six==1.15.0               # via lizard-auth-server, pyjwkest, python-dateutil
text-unidecode==1.3       # via faker
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.25.11          # via requests
zipp==3.3.1               # via importlib-metadata
